BC-461: Added zap.context
