+++ .github/workflows/test.yml
@@ -204,7 +204,7 @@ jobs:
           target: 'http://localhost:8080/?providerAccountNumber=0P322F'
           rules_file_name: '.zap/rules.tsv'
           allow_issue_writing: false
-          cmd_options: '-j' # https://www.zaproxy.org/docs/docker/full-scan/#usage
+          cmd_options: '-j -z "-config spider.maxDepth=10" -n zap.context' # https://www.zaproxy.org/docs/docker/full-scan/#usage
       - name: Upload logs
         uses: actions/upload-artifact@v6.0.0
         with:
