+++ src/test/java/uk/gov/justice/laa/amend/claim/service/MaintenanceServiceTest.java
@@ -32,7 +32,7 @@ public class MaintenanceServiceTest {
     private HttpServletRequest request;
 
     @InjectMocks
-    private MaintenanceService service;
+    private ConfigurableMaintenanceService service;
 
     @Captor
     private ArgumentCaptor<Path> pathCaptor;
